Paging is now enabled!
I still had a couple of invalid type warnings to fix, but the big issue was
getting the correct value for the pointers to the PD entries. As paging is
not yet enabled when they need to be set, they must be physical addresses and
not virtual.

This should be the end of this physical to virtual instruction offset garbage
from now on. I should just have to set up the correct page tables and be
done with it. Keywords: ``should'' and ``just''.